Delete gdb.cc with casting methods
gdb (since forever, it appears) supported 'set print object' to
"identify the actual (derived) type of the object rather than the
declared type". The setting can be enabled either globally or via
the -o/--object flag to the print command, and works fine at least
for a 'Grob *g' being an 'Item *'.