Write input file for Ghostscript as UTF-8
Since commit 017927b4d6 ("Unify calling convention for command-line
and API GS use"), we pass the input and output file names via a
temporary file. That needs to be written as UTF-8 so that file names
with special characters work. As a side effect, stop re-encoding
the command line arguments to Ghostscript because these are purely
ASCII now.

Fixes #6089