Allow \alternative {...} inside a repeat body
Also allow assigning a group of alternatives to a variable and using
it later as \alternative \var.

The "More alternatives than repeats" warning now refers to the source
location of both the \repeat and the \alternative.

There is a new type of music, SequentialAlternativeMusic, and a new
iterator, Alternative_sequence_iterator.  This iterator takes over most
of what the Volta_repeat_iterator used to do.

Known issue: End-justified alternatives are not yet supported inside the
repeat body: the end-repeat bar lines are incorrect.  The legacy syntax
must be used in that case.