Extend test figured-bass-continuation-center.ly
Add more combinations to test the behavior on multiple groups. Note
that the logic in Figured_bass_engraver is split between the end of
process_music and center_repeated_continuations, but at least this
change makes it impossible to delete code from either without seeing
the result in the test.