Introduce 'debug-gc-lifetimes option
The debug-gc-assert-parsed-dead feature has been documented, but it is
really an internal mechanism (it only has effect during garbage
collection, and has to be coupled with calls to ly:parsed-undead-list!
in order to do something useful.

Instead, introduce the debug-gc-object-lifetimes option, which
controls the entire check (ie. the extra GC call, and the printing of
objects afterwards). For backward compatibility, this feature defaults
to #t.