Reset modules after each file with Guile 2.x
Guile 2.x attaches all top-level modules to a root module named '()
(not be confused with the-root-module aka (guile)). They are stored
in the submodules field, which makes them stay alive even after all
other references are gone.
If a module doesn't have a name, it is assigned a generated symbol
in the accessor (module-name). Afterwards, the module is passed to
(nested-define-module!) with '() as its parent. For LilyPond, that
accessor is used from (module-export-all!), which is called right
after a module is created in ly_make_module.

Having all modules stay alive is problematic because it also keeps
all defined variables around. This includes those from .ly files,
which are leaked across sessions. Furthermore it also prevents most
Font_metrics from being collected, so Pango fonts are retained in
spite of the recent fixes.

Fixes #6081