Make -ddebug-gc-object-lifetimes more robust
This feature was introduced in commit 7785ac38 (Nov 3, 2006) as a
mechanism to detect memory leaks: when an object is found alive when
we are not parsing a file (eg. a Grob), print out a warning,
indicating that some code is retaining references across .ly files
that it should not.

The current implementation has some downsides:

* it prints just a single offending object of each type, which limits
  its usefulness in diagnostics

* hiding the initial file behind the `undead` smob may also hide objects
  that are incorrectly added to mutable data structures reachable from
  the initial initialization.

Instead, store *all* 'forbidden' objects during mark. Report if they
happen after the first file.  Store them in hash table, to avoid
reporting the same object twice.

Testing with GUILEv2 shows that the mark function is called for
to-be-deleted objects too, so this mechanism can't be made to work
with GUILEv2.

Adds a scheme-zombies.ly file to verify that the functionality is
working

The value of the debug-gc-object-lifetimes may be an integer, denoting
the maximum number of warnings to print.