Remove STEPMAKE_FLEXLEXER_LOCATION
Since commit 03a98b33f1 ("Stop copying FlexLexer.h"), the variable
is only used to depend on the header. This is atypical because it
is not our own. Moreover the check breaks with Autoconf 2.70 with
"configure: line 6770: conftest.cc: command not found", probably
because $ac_cpp is not set anymore.