scm: Replace with-location by explicit calls
To make ly-syntax-constructors.scm compilable into bytecode, it would
be required to have direct access to the %location fluid. That defeats
the purpose that there are no public setters, so solve the problem by
calls to (ly:set-origin!) with an explicit location.
I've verified that a \addlyrics command still results in the same value
of 'origin for the created ContextSpeccedMusic and LyricCombineMusic.

This effectively reverts the following two commits from Issue 4474:
873f9fd277 Issue 4474/5: Provide with-location macro
8499cabe34 Issue 4474/6: Combine lyric-combine and lyric-combine-music