common-macros.itexi: Simplify `@sourceimage` macro.
Use only three arguments and make the `alttext` argument identical to the
filename (without extension).

Update all callers by running the following script.

  for i in `find . -name '*.itexi'`; do
    sed -i -e 's/\(@sourceimage{[^}]*,[^}]*,[^}]*\),[^}]*}/\1}/' $i
  done

  for i in `find . -name '*.itely'`; do
    sed -i -e 's/\(@sourceimage{[^}]*,[^}]*,[^}]*\),[^}]*}/\1}/' $i
  done

Note that some macros calls had 'png' or 'jpg' as an `alttext` value, which
doesn't make sense.  This is due to older modifications of the
`@sourceimage` macro: Previous versions also passed an extension as an
argument, which is no longer the case, and the calls weren't correctly
updated after the change.