Avoid transparent bar lines in ancient staves
Instead of using transparent "|" bar lines which still take up space,
let the Bar_engraver use a different default bar line ("" in this case)
if defaultBarType is overridden in a lower context.  This closes #6098.

Also, override the default line thickness of the BarLine grob so that
section ("||") and final ("|.") bar lines may be used and are rendered
as a finalis sign.