Initialize session from a dedicated parser
The session mechanism reads the initialization .ly files, to setup
variable assignments to replay for later files.  This mechanism
inherently retains the module associated with the parser for
initializing the session.

Before, if the first .ly file assigned something to an identifier, the value
would be retained, leading to "parsed-object is dead" warnings.

By executing the initialization in a separate parser, we avoid the
warnings. It also ensures that the first file is processed in the same
way as later files.