Recognize empty bar lines by shape, not glyph name
is_loose_column() used to look for an empty glyph name, but now it looks
for a bar line with zero width.  This closes #6097.

A side effect of this change is visible in regression test
staff-online-symbol-absence.ly, seemingly due to issue #4084, "Hidden
BarLine during note yields programming error".  I considered changing
note durations to avoid having a note straddling measures, but the
author's choice of \time 5/8 suggests that it was intended.