IR: Improve cross-references in PDF output.
This commit changes cross-references in the IR from

  This context creates the following layout object(s): Section 3.1.13
  [BarNumber], page 369, Section 3.1.23 [BreakAlignGroup], page 379, ...

to the much more readable

  This context creates the following layout object(s): BarNumber (page 353),
  BreakAlignGroup (page 363), ...

It also reduces the length of the IR by a bit more than approx. 3% (which
is currently 23 pages).

Heart of the change is a new macro `@iref` that the documentation generator
now uses instead of `@ref`.  For non-PDF output it yields the same results
as before; for PDF, some new TeX macros are provided to get the desired
output.