Add bar lines and improve test coverage
Add variants of `[|:` and `S.|:` annotated with `-||` to parallel
`.|:-||`.

Add bar lines for coincident Fine and in-staff segno.

Replace regression test double-repeat-default-volta with
bar-line-define-repeat.  The new test adds coverage for overriding
startRepeatType and endRepeatType and using user-defined bar lines.

Add tests bar-line-built-in-*.ly using the \bar command with all
built-in repeat and segno bar types at the beginning, middle, and end of
a system.  Delete a few tests that covered a subset of these.