\section command
The \section command places a section break into the music.  In this
commit, a section break creates a thin double bar line, but this is
intended as a semantic command that could trigger additional behavior
from engravers or performers in the future.

In contrast to \bar "||", \section does not override repeat bar lines,
but triggers the use of annotated bar types that allow the section bar
line to appear at line breaks.

Ideas for the future:

  * \sectionLabel to add text marks like INTRO, REFRAIN, CODA
  * line breaking and spacing might treat \section specially