Add public function break-alignment-list
There is no convenient way to specify different alignment values for a
grob depending on where it appears in a line. Add a new public scheme
function that allows users to specify different alignment values for
beginning, middle, or end of line. Add a test for it using bar numbers.