Ignore baseline-skip in the triangle markup command
The \triangle markup command produces glyphs that are too large at large
font sizes, and too small at small font sizes. It also causes LilyPond
to abort when the baseline-skip value is 0 - producing no PDF but
leaving behind temporary files.

Remove baseline-skip from the calculation of the triangle and update the
existing multiplier to a value that produces triangles that scale more
proportionately with font size. This also circumvents the problem with
LilyPond aborting. Add a reg test.

Closes #6096