Fix event-listener.ly when input file name contains spaces
In routine `filename-from-staffname`, use ly:parser-output-name to
return the same string as is used to build eg the `output.pdf`
name.

This correctly handles spaces in the filename, whereas the previous
scheme which backed up from the end of a command line argument to
the last space yielded unexpected names for the `.notes` file.