Error out on unrecognised percussion signs
If a user adds an undefined sign to the drumPitchTable (i.e. one that
isn't in the default-script-alist), Lilypond will crash with a failed
assertion when it tries to find the stencil for it.

Check for this problem and fail with a useful error message instead.

Closes #6091

(cherry picked from commit cbc0935403ab50136d848181201d1505699017d1)