Preserve the order of post events when they are wrapped
Commit 44956f9b3ff (unintentionally?) reversed the order of post events
when they are wrapped. Although it seems we make no guarantees about
ordering, it can cause unexpected output because the order is the
reverse of when the events are not wrapped.

Un-reverse the order of elements in post-event-wrappers so that they
appear in the same order as if they had not been wrapped. Add a regtest
to catch this in future.

(cherry picked from commit 32b1b7fa7834a9e31a8b00a3969a44b9e6721d7f,
changed \version in the regtest to 2.22.1)