Let \incipit set self-alignment-X only for first InstrumentName
Before, it was a general setting and couldn't be customized for
the shortInstrumentName any longer.
Also, extend regtest incipit.ly to actually cover \incipit and
check for shortInstrumentName's self-alignment-X setting.
Some reformating there.

Actually there is still some difference in aligning instrumentName
and shortInstrumentName. I suspect it is related to padding. Worth
further investigation, but beyond the scope of this patch.

Closes #6108