lilypond-book: (Better) support of inline images
This commit introduces a new snippet option, 'inline'.  For LaTeX and
docbook output, this option is redundant; inline images are already
supported.  For HTML and texinfo output, however, it is necessary.  Note
that `lilypond-book`'s documentation for HTML output incorrectly stated that
inline images were supported; it seems that this capability was available in
the past but accidentally removed.

The following changes and fixes are applied.

* book_html.py: Correctly support the 'quote' snippet option.  Up to now,
  it only tried to put verbatim text (using `<pre>`) into a `blockquote`
  environment.  Now the snippet image is added to `blockquote`, too.

  Handle 'inline' snippet option by inserting `<p>` (and `</p>`)
  conditionally.

* book_texinfo.py: Handle 'inline' snippet option by inserting empty lines
  before and after the image conditionally.  Similarly, insert `<p>` and
  `</p>` (for `texi2html`) conditionally.

* book_snippets.py: Don't apply left padding if 'inline' option is set.