Derive all StaffGroup-like contexts from StaffGroup
Previously, there were many inconsistencies between the context types
StaffGroup, ChoirStaff, GrandStaff and PianoStaff.  They had differences
in their lists of accepted contexts, and setting localAlterations or not.
This lets ChoirStaff and GrandStaff inherit from StaffGroup (PianoStaff
still inherits from GrandStaff) to keep all of them in sync.

StaffGroup gains a bunch of \accepts, so it accepts everything also
accepted by Score.

This fixes a problem occurring when a TabStaff was included in a ChoirStaff.
The latter did not accept it, causing a StaffGroup to be inserted in between,
with its own bracket.

Closes: #6094