Introduce the centered-text-interface
This defines a new grob interface with a stencil callback,
centered-text-interface::print.  It serves to print any
kind of text between two columns, nicely aligned with
spacing-pair.

The larger goal of this change is implementing centered
measure numbers (done in a later commit).

Another perspective for the future is a simple CenteredTextSpanner,
which is a popular feature request.