Give MetronomeMark a higher outside-staff-priority
This changes MetronomeMark's outside-staff-priority from 1000 to 1300.
The only possible change in ordering is with piano pedal spanners,
which have a priority of 1000. It's highly unlikely to find a metronome
mark and a piano pedal at the same position, so this should not impact
user scores.

This change is done in the perspective of implementing measure-centered
bar numbers: they should be farther from the staff than piano pedals,
but closer than metronome marks.