Implement centered bar numbers
The Bar_number_engraver reads a new context property, centerBarNumbers.
If that is true, it creates CenteredBarNumber grobs instead of the
regular BarNumber grobs.

CenteredBarNumber is a new grob type.  It is a Spanner, on the contrary
to BarNumber, an Item, and has its own set of properties.

Centered bar numbers, as opposed to regular bar numbers, are vertically
aligned on the same position.   This is achieved through making them
children of a CenteredBarNumberLineSpanner, similar to DynamicLineSpanner
and the ...PedalLineSpanner types.

Three grob interfaces are added.

- The bar-number-interface groups all grobs related in some way to bar
  numbers.
- The centered-bar-number-interface is specific to CenteredBarNumber and
  used by the Centered_bar_number_align_engraver.
- The centered-bar-number-line-spanner-interface is specific to
  CenteredBarNumberLineSpanner.  It may be used to filter out those grobs
  from the "actual" bar numbers (this is consistent with dynamic-interface
  and dynamic-line-spanner-interface).