Let the measure counter account for compressed MM rests
The Measure_counter_engraver reads currentBarNumber instead of
maintaining a local count variable.  It uses barNumberFormatter
to make a markup out of it.

When it detects that more than one measure is spanned, it prints
a hyphen-separated measure range.

Using barNumberFormatter also has the nice effect that alternative
numbers are honored.

Closes #6124.