Docs: Avoid broken links to translated PDFs
This isn't fully correct either since it now always links to the
English PDF on the web page generated with the releases. However,
the really important links on the front page of lilypond.org are
correct (or at least they can be) since it's passing the full URL
during the website build.

Solves the rest of #6037