lilypond-book: Only append the current directory
This allows include directories to shadow files from the current
directory, which is needed for the documentation build: We must
not change to $(outdir) as done by commit d3a78ea9aa or the .dep
files list dependencies for the wrong target. To still find the
translated snippets, $(outdir) must be ordered ealier than $(srcdir)
which is the current directory for in-tree builds.