Revert DynamicText.extra-spacing-width in ChoirStaff context
Dynamics are usually allowed to cross columns.  In StaffGroup,
GrandStaff and PianoStaff, this is undesirable because they
can collide with span bars.

This is not the case in ChoirStaff, so unset the
DynamicText.extra-spacing-width property to allow for tighter
spacing.

This fixes a regression introduced by commit e31b0b75.  It would
be nice to find a way to prevent dynamics from crossing span
bars exclusively, allowing for tight spacing mid-measure, in
StaffGroup context, but for now this at least avoids worsening
the spacing in user scores.