Let markup interpretation fall back to font-defaults for alteration glyphs
A new markup command, \accidental, is introduced.  It takes the alteration
(as an exact rational) and prints the corresponding glyph, which is
chosen from the alteration-glyph-name-alist property, defaulting to the
value in font-defaults added in the previous commit.

The commands \markup \sharp, \markup \flat, etc. are refactored to
use \markup \accidental, which lets them support the
alteration-glyph-name-alist property instead of hardcoding the default.

Note names and chord names now use \markup \accidental for their
text.  This allows for overriding their alteration-glyph-name-alist,
and sets the default to the one in font-defaults.

Include files that define a style with non-standard accidentals
(such as makam.ly) no longer have

(set! standard-alteration-glyph-name-alist ...)

This line was causing session bleedover in documentation builds
(https://lists.gnu.org/archive/html/lilypond-devel/2021-05/msg00001.html).

Another benefit is bringing consistency: all grobs that print accidentals,
including all those that interpret markups (TextScript, InstrumentName,
etc.) now support the alteration-glyph-name-alist property.  The likes
of makam.ly now set systematic defaults, which work for all kinds
of grobs.

How to set the default globally is not documented: the next
commit introduces a user interface for that, with a context
property.