Copy images as part of the install-info make target
By default, install-info uses a symlink to html documentation for
included images. This results in broken links unless the user also runs
make install-WWW, which is unlikely since there is already a target for
exactly that combination: install-doc.

Update install-info to copy images, but keep the old behaviour when
install-doc is run (because the symlink will work in that case and there
is no need to copy duplicate images).

Closes #6110