Require createSpacing = ##t to create LeftEdge
Before this change, a BarLine originating in a Dynamics context could
trigger the creation of a LeftEdge using the spacing settings visible
in the Dynamics context.  It is preferable to create a LeftEdge using
the spacing settings visible in a Staff context.

This change defers creation of a LeftEdge until it can be done in a
context with createSpacing = ##t in view: among the built-in contexts,
this means Staff, RhythmicStaff, or one derived from either.

Closes #6134.