lilypond-book: Remove snippet map
Support was broken ever since commit 38e1cad48d ("Fix the linking
of files from lilypond by lilypond-book.") prepended the current
directory as relative path to the list of includes in 2012. This
made LilyPond find the snippets as `./??/lily-*.ly' and the whole
mapping did not work. Now that commit 06cdcec72f ("lilypond-book:
Convert to absolute paths earlier") uses absolute paths and the
--lily-output-dir is not in the list of include paths, the mapping
came back to life and it becomes apparent why nobody missed it:

 * It's redundant since --use-source-file-names will replace the
   input file name with the original source file name.
 * Since the snippet map is added as a regular entry to the list
   of input files, it is subject to processing with multiple jobs.
   In particular this means:
     - without the argument -djob-count or a requested number of
       jobs greater than the number of input files (which also
       results in no jobs being forked), things might work;
     - if jobs are forked (and be it only one), the list of input
       files gets reversed and the map comes last after all files
       have been processed; and
     - even if that was fixed, only one job gets to see the map.