Fix balloons on grobs with empty pure Y extent
This is done by introducing a new function, robust_relative_pure_y_extent,
which is the pure equivalent of robust_relative_extent.  When the extent
is empty, it returns a single-point interval rather than the empty interval.
This is accurate for balloons, because they use robust_relative_extent
in the unpure stencil.

Closes #5306.