Fix balloons on grobs with empty pure Y extent
This is done by introducing a new function, robust_relative_pure_y_extent,
which is the pure equivalent of robust_relative_extent.  When the extent
is empty, it returns a single-point interval rather than the empty interval.
This is accurate for balloons, because they use robust_relative_extent
in the unpure stencil.

Closes #5306.

(cherry picked from commit 10e8952de0494a65323e9d9b2ecef129c7f7f8ca,
changed \version in the regtest to 2.22.2)