Reimplement \ottava with a stream event
This has several benefits over the previous approach with an
elements-callback of OttavaMusic.  For one thing, the detection
of a new ottava is more reliable.  It was based on the labels,
comparing them with eq?.  Guile is free to cache short strings,
making (eq? "8" "8") true, which caused a bug where consecutive
\ottava 1 and \ottava -1 were merged together (the second was
ignored).

For another, because everything is baked in the engraver, it now
works to move it to Voice context to obtain per-voice ottavas.
This used to require workarounds based on middleCOffset.

Closes #6141, #1473.