Reduce unsmobbing during break substitution
substitute_grob (SCM, Grob *) needed to figure out what kind of scheme
value it had been given, but that knowledge is present a little higher
in the call chain.  Split it into two functions for the two possible
cases.  Use templates to preserve the type information through a few
levels of callers.