Refactor polygon stencils
Previously, polygons stencils could be created in Scheme using
ly:make-stencil.  There was ly:round-filled-polygon, making a
polygon with round edges, but always filled.

This extends ly:round-filled-polygon, renamed to ly:round-polygon,
with an optional boolean parameter controlling whether the polygon
is filled.  A markup command is added as a user interface and for
future uses in displaying control points.  The extroversion parameter
now defaults to 0 rather than -1 because this makes more sense from
the user point of view.

The implementation of \markup \triangle is slightly simplified.
As side benefits, it now scales the thickness correctly by
line-thickness, honors the extroversion property, and takes the
thickness into account in the extents.