Unconditionally warn about cyclic dependencies
Check for cyclic dependencies and warn about them even in non-debug
builds.  This is useful even for users: if LilyPond has a bug that lies
in a cyclic dependency, it is better to be informed about it so you
have a change to report it to the mailing lists rather than have
LilyPond behave surprisingly without any indication why.

It is also useful for those who extend LilyPond in Scheme, who
(from experience) can hit cyclic dependencies relatively easily.

Finally, it is useful for developers, because we might get more
bug reports.

This is also consistent with the code for pure heights, which
always warns about dependency cycles.

As a result, the check in get_grob_direction can be removed
since get_property now does the job.