Simplify linear_combination (CENTER)
Add Drul_array<Real>::average () to replace a couple of calls to
Drul_array<Real>::linear_combination (CENTER).  The remaining call with
a variable argument (in Beam::calc_stem_y) is removed by working with an
Interval instead.

Move the implementation of linear_combination () to Interval_t<Real>.

Specialize Interval_t<Real>::center () so that it can be called in the
the same situations as linear_combination (CENTER) without failing an
assertion.  Replace calls to the latter by calls to the former.