Let Time_signature_performer create proper metronome data
Previously, Time_signature_performer put a hardwired value of 18,
corresponding to a rather strange value of 3/16 as the "metronome"
beat into the MIDI time signature.  Many interpreters of the MIDI file
ignore this information but getting it right does not seem amiss.

This commit uses the "beatMoment" value times the gcd of the numbers
in a possible "beatStructure" to derive the principal metronome unit.
That has a good probability of being correct.

The metronome unit can be verified (for example) using `lilymidi
--pretty` on a MIDI file containing time signature data.