Drul_array: add front() and back()
The constexpr_assert() function added recently [1] does not work in
versions of GCC before 5.  GUB is still using 4.9.4.

Revert to assert() in Drul_array::at() and operator[].  The value of
these checks is low but not zero.

Add front() and back() methods for unchecked access that can't fail.
Their const overloads are also constexpr.

[1] commit 465213cf2c898ff649d423c207b11e737b6d891e