Allow engravers to decide about grob classes
This is a conceptual shift: the class of a grob (Item, Spanner,
Paper_column, System) is no longer always determined by its
type (NoteHead, ...).  Engravers have the option to decide what
class to create the grob with.  This means that the same grob
definition in define-grobs.scm can be used to make both items
and spanners.

The allowed classes are stored in the meta.classes field of
grob definitions.  If there is just one class in the list,
ly:engraver-make-grob can be used, and it chooses this
class.  By contrast, if there are two or more entries in
meta.classes, engravers must specify the class by using
either ly:engraver-make-item or ly:engraver-make-spanner
instead of ly:engraver-make-grob.  The C++ macros make_item,
make_spanner and make_column work the same internally
as ly:engraver-make-item and ly:engraver-make-spanner:
create a grob with the class corresponding to the name
of the macro, and check that this class does belong to
the meta.classes property of the newly created grob.
(Note that ly:engraver-make-column is not provided,
since creating paper columns from custom engravers is
not supported.)

For simplicity and compatibility, grob definitions may contain a
meta.class field.  It should be just a symbol representing one of
the available classes.  In this case, the meta.classes field is
set to a one-element list containing this symbol.  This is done
upon completizing the grob definition.  In further commits, the
possibility of meta.classes introduced by this commit gets used
by a some grobs, but the wide majority still just declares meta.class.

This change implies that the interface that characterizes a class
(e.g., item-interface) is no longer added when completizing grob
definitions during initialization, but when instantiating grobs.
This requires an adaptation to the acknowledging cycle since it
relied on the assumption that the same grob name would lead to the
same set of interfaces. Now, the hash table entry is indexed by
the (grob-name . grob-class) pair. Thus, class_name () needs be
public for grob classes, and hashq-create-handle! must be replaced
with hash-create-handle!.

The documentation generation scripts are updated to display the
allowable classes now that they are no longer implied by the
set of interfaces found in the completized grob definition.

\alterBroken can no longer detect just from the name of the
grob that it is not a spanner. The warning is moved to
value-for-spanner-piece.