Parentheses as sticky grobs
This makes parenthesizing spanners supported.

The refpoint calculation in the stencil callback did not include the
friends.  This has worked so far because the parent of Parentheses was a
PaperColumn, which is deep enough to be in the lineage of accidentals
and dots.  Because sticky grobs have their host as X parent, this commit
requires deepening the reference point to include friends. (This change
is not separate because ly:grob-common-refpoint-of-array insists to
include the grob itself in the refpoint calculation too, leading to
problems with parenthesizing non-musical grobs such as breathing signs
without making the parenthesized grob parent to the parentheses.)

Closes #4565, #2646.