Split mark events into two classes
One class represents a true rehearsal mark and the other represents the
(ab)use of a RehearsalMark grob to position arbitrary markup.

For true rehearsal marks, limit the 'label property to an index.  Set
'text directly for ad-hoc marks.

Mark_engraver still listens for a common mark event, but enhancements
are planned.

This is a more conservative version of the patch in Issue #5301.