Base accidental rules on global timing, not measure position
measurePosition stops increasing during a cadenza.  It is thus
unreliable for determining automatic accidentals, which
(unlike, e.g., automatic beams) should only depend on measure
boundaries, not on the measures' inner structure.  Issue #6153
was caused by incorrect detection of consecutive notes due to
this.  Use the context's current moment instead.  This also
fixes a similar problem with partials: the function was called
with the bare measurePosition, but the locations stored in
localAlterations had a version that is normalized by adding
a multiple of the measure length to make it positive.

The current moment is not passed to the functions as the measure
position used to be since it can be easily retrieved with
ly:context-current-moment.

Closes #6153.