Clean up Interval_t<T> code relating to "infinity"
Instead of requiring all specializations of Interval_t<T> to provide a
method returning a maximum value for T, use a traits type with a default
that relies on T itself.  Thus, the traits type does not need to be
specialized for Rational, Moment, or any other types we control.

Also, deal with the minimum value on its own rather than assuming that
the negative of the maximum value is in range.  That is not the case for
vsize.

Make the constructors for an empty interval and the longest interval
constexpr.

Merge TEST (Interval_test, unite) into TEST (Interval_math_test<Mint>...

Add tests of add_point().

Test is_empty() for more types.

Closes #6176.