ContextChange command can affect a tagged scope
Add an internal option to apply a ContextChange command to iterators in
a specified musical scope.   That scope is the nearest enclosing music
tagged with the tag given in the command's 'change-tag property.

This will allow expressing \autoChange and \partCombine using normal
sequential music instead of an ad-hoc list of times and names, and that
will allow transforming it in some new ways (e.g. with \scaleDurations)
without additional special handling.

    \tag X <<
      { ... skips and changes to apply within scope X ... }
      { ... part notation ... }
    >>

The changes to \autoChange and \partCombine follow in separate commits.

There is no user interface for creating ContextChange commands that
specify a tag.  The current goal is just to make the internal
representation easier to work with.  An important decision about the
nature of simultaneous music would need to be made before promoting the
above pattern for general use: are elements of << >> interpreted in
source order at a given timestep?  The pattern depends on that, so it
would either be necessary to impose that constraint on simultaneous
music generally, or to provide a way to request it in a specific case.