Add a time-based version of \parenthesize
\parenthesize can now work either on a music event or on a grob path,
just like \footnote.  This makes it possible to parenthesize grobs
that do not find a cause in events, such as clefs.

The implementation just applies a propertyTweak.  This preserves the
behavior of parenthesizing every element of a chord.

An unfortunate side effect is that something like
\lyricmode { \parenthesize aaah } no longer works since "aaah" is taken
as a grob path.  This is easily worked around using
\lyricmode { \parenthesize LyricText aaah }.