Music_iterator::preorder_walk
music_iterator->preorder_walk (f) calls function f on music_iterator and
then on all its children.

substitute_context () now acts non-recursively and preorder_walk () is
used for the recursion.

This separation makes the code more flexible for the future and easier
to navigate in the present.  Only the subclasses that need to change the
recursion override preorder_walk () and only the subclasses that deal
with references to additional contexts override substitute_context ().