Add set of scripts to build static binaries on Linux
This initial commit implements the necessary infrastructure to
natively build all of LilyPond's dependencies and statically link
them into a single lilypond binary that only depends on fundamental
libraries, such as glibc on Linux. The result is packaged into an
archive that can be extracted anywhere and used without installation.
That archive also includes interpreters for the scripts and places
wrappers in the bin/ directory for convenient execution.

The added README.md provides details on some key design decisions,
with the major points being:
 * Build natively, avoid cross-compilation.
 * Use static builds of dependencies, no shared libraries.
 * Be specific to LilyPond, no generic package manager.
 * Use tarballs for sources, no download from git.
 * Keep it simple, stupid.

In the near future, I plan to extend the scripts to also produce
binaries for FreeBSD, macOS, and Windows via mingw.

(The system presented here is a complete rewrite of the previously
shared proof-of-concept based on shell scripts, which already has
support for the other platforms mentioned above. It remains available
at https://github.com/hahnjo/lilypond-binaries/.)