recording-group-emulate records context end time
The output of recording-group-emulate now has a final element telling
the time the context was removed.  This element has an empty event list.

The global context needs to check for removable contexts after quitting
the top-level iterator so that the expected RemoveContext events are
sent.

Part combiner code in make-voice-states no longer has to generate its
own final entry.

Some code related to quoting previously took a shortcut to tell whether
the output of recording-group-emulate held any events.  The shortcut is
no longer possible now that there is always an element for the end
moment.  The code now checks for a non-empty event list in the entry for
the first moment.