Fix balloons on cross-staff grobs
Since balloons have a height that depends on their parent, they must
marked cross-staff when their parent is cross-staff (at least when the
Balloon_engraver is not in Score context, as if it is, the grobs won't
participate in their parent's axis group in the current implementation).

For now, BalloonText is the only sticky grob with a cross-staff
callback.  Footnote, ControlPoint and ControlPolygon do not need it
because they have extents set to #f (but testing coverage is added
nevertheless).  The case of Parentheses is more tricky because handling
it correctly would requires something like a version of
ly:relative-group-extent that does take cross-staff grobs into account
(issue #4892); for now, it is left as-is.

Closes #6161.